[REF] refactor the code in order to split it in several file
