+++ mail_usability/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import mail
-from . import tools
-from . import mail_template
-from . import mail_message
+from . import models
+from . import wizard
