+++ mail_usability/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from . import mail
+from . import tools
+from . import mail_template
+from . import mail_message
+from . import res_partner
