+++ mail_usability/wizard/email_template_preview_view.xml
@@ -1,23 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  © 2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
-  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
--->
-
+<?xml version="1.0" encoding="UTF-8"?>
 <odoo>
-
-<record id="view_mail_tree" model="ir.ui.view">
-    <field name="name">mail_usability.mail.tree</field>
-    <field name="model">mail.mail</field>
-    <field name="inherit_id" ref="mail.view_mail_tree"/>
-    <field name="arch" type="xml">
-        <field name="email_from" position="replace"/>
-        <field name="date" position="after">
-            <field name="email_from"/>
-            <field name="email_to"/>
-        </field>
-    </field>
-</record>
 
 <record id="email_template_preview_form" model="ir.ui.view">
     <field name="model">email_template.preview</field>
@@ -38,4 +20,5 @@
         </footer>
     </field>
 </record>
+
 </odoo>
